[feature] RestDocsIntegrationTest 추가
설명:
- RestDocs를 작성할 때 해당 클래스를 상속받아 작성할 수 있도록 추상클래스를 작성했습니다.
